Create Agenda Management (3rd Feature)
This code creates a simple user interface for the Agenda Management feature. It includes an input field for adding new agenda items, a button to add the item to the agenda, and a FlatList to display the list of agenda items. When the "Add" button is pressed, the new agenda item is added to the list of items, which is stored in the agendaItems state. The list of items is displayed using a FlatList component, with each item being rendered as a separate view.