Create Communication and Collaboration(4th Feature)
This code provides a basic implementation of the communication and collaboration feature for the virtual event planning and management app, allowing users to send and receive messages within the app. The messages are stored in an array and displayed using a ScrollView component. The user can type their message in a TextInput component, and the "Send" button sends the message and adds it to the list of messages.